Boundary conditions and forcings

CryoGrid.Physics.Boundaries.ConstantBC — Type

ConstantBC{S,T} <: BoundaryProcess

Constant boundary condition (of any type/unit) specified by value.

CryoGrid.Physics.Boundaries.Damping — Type

Damping{D,K} <: BoundaryEffect

Generic implementation of bulk conductive damping at the boundary.

CryoGrid.Physics.Boundaries.Forcing — Type

  Forcing{T,N}

Abstract type representing a generic external boundary condition (i.e. "forcing").

CryoGrid.Physics.Boundaries.PeriodicBC — Type

PeriodicBC{S,T} <: BoundaryProcess

Periodic boundary condition (of any type/unit) specified by period, amplitude, and phaseshift.

CryoGrid.Physics.Boundaries.TimeSeriesForcing — Type

  TimeSeriesForcing{T,A,I}

Forcing provided by a discrete time series of data.

CryoGrid.Physics.Boundaries.TimeSeriesForcing — Method

Get interpolated forcing value at t seconds from t0.