Heat conduction

CryoGrid.Physics.HeatConduction.FreeWater — Method

(fc::FreeWater)(sub::SubSurface, heat::Heat{FreeWater,Enthalpy}, state)

Implementation of "free water" freeze curve for any subsurface layer. Assumes that 'state' contains at least temperature (T), enthalpy (H), heat capacity (C), total water content (θw), and liquid water content (θl).

CryoGrid.Physics.HeatConduction.enthalpy — Method

enthalpy(T, C, L, θ) = T*C + L*θ

Discrete enthalpy function on temperature, heat capacity, specific latent heat of fusion, and liquid water content.

CryoGrid.Physics.HeatConduction.heatcapacity! — Method

heatcapacity!(sub::SubSurface, heat::Heat, state)

Computes the heat capacity for the given layer from the current state and stores the result in-place in the state variable C.

CryoGrid.Physics.HeatConduction.heatcapacity — Method

heatcapacity(sub::SubSurface, heat::Heat, state, i)

Computes the heat capacity at grid cell i for the given layer from the current state.

CryoGrid.Physics.HeatConduction.heatconduction! — Method

heatconduction!(∂H,T,ΔT,k,Δk)

1-D heat conduction/diffusion given T, k, and their deltas. Resulting enthalpy gradient is stored in ∂H. Note that this function does not perform bounds checking. It is up to the user to ensure that all variables are arrays of the correct length.

CryoGrid.Physics.HeatConduction.liquidwater — Method

liquidwater(sub::SubSurface, heat::Heat, state)
liquidwater(sub::SubSurface, heat::Heat, state, i)

Retrieves the liquid water content for the given layer at grid cell i, if specified. Defaults to using the scalar total water content defined on layer sub.

CryoGrid.Physics.HeatConduction.thermalconductivity! — Method

thermalconductivity!(sub::SubSurface, heat::Heat, state)

Computes the thermal conductivity for the given layer from the current state and stores the result in-place in the state variable C.

CryoGrid.Physics.HeatConduction.thermalconductivity — Method

thermalconductivity(sub::SubSurface, heat::Heat, state, i)

Computes the thrmal conductivity at grid cell i for the given layer from the current state.

CryoGrid.Physics.HeatConduction.totalwater — Method

totalwater(sub::SubSurface, heat::Heat, state)
totalwater(sub::SubSurface, heat::Heat, state, i)

Retrieves the total water content for the given layer at grid cell i, if specified. Defaults to using the scalar total water content defined on layer sub.