Hydrology

CryoGrid.Physics.Hydrology.BucketScheme — Type

BucketScheme{Tfc} <: WaterFlow

"Bucket" water scheme for downward advective flow due to gravity.

CryoGrid.Physics.Hydrology.DampedET — Type

DampedET{Tftr,Tfev,Tdtr,Tdev}

Corresponds to evapotranspiration scheme 2 described in section 2.2.4 of Westermann et al. (2022).

CryoGrid.Physics.Hydrology.Evapotranspiration — Type

Evapotranspiration

Base type for parameterizations of evapotranspiration (ET).

CryoGrid.Physics.Hydrology.HydraulicProperties — Type

HydraulicProperties

Material hydraulic properties.

CryoGrid.Physics.Hydrology.Rainfall — Type

Rainfall{Train<:Forcing{u"m/s"}} <: BoundaryProcess{WaterBalance}

Basic rainfall boundary condition for WaterBalance which simply invokes the given precipitation forcing at the current time t.

CryoGrid.Physics.Hydrology.WaterBalance — Type

WaterBalance{TFlow<:WaterFlow,TET<:Union{Nothing,Evapotranspiration},Tdt,Tsp,TProp} <: CryoGrid.SubSurfaceProcess

Represents subsurface water transport processes.

CryoGrid.Physics.Hydrology.WaterBalanceProperties — Type

WaterBalanceProperties

Numerical constants shared across water balance implementations.