Interface

CryoGrid.Interface.AbstractParameterization — Type

AbstractParameterization

Base type for representing parameterizations.

CryoGrid.Interface.Bottom — Type

Bottom <: Layer

Generic "bottom" layer that marks the lower boundary of the subsurface grid.

CryoGrid.Interface.BoundaryProcess — Type

BoundaryProcess

Abstract base type for boundary processes, i.e. processes that operate at the boundaries of the subsurface. A BoundaryProcess represents the boundary conditions of one or more SubSurfaceProcesses but may include its own diagnostic (or even prognostic) variables, if necessary.

CryoGrid.Interface.BoundaryStyle — Type

Trait that specifies the "style" or kind of boundary condition. This can be used to write generic implementations of interact! that are (relatively) agnostic to specific implementations of BoundaryProcess. A good example of this can be found in HeatConduction.boundaryflux.

CryoGrid.Interface.BoundaryStyle — Method

BoundaryStyle(::Type{T})

Can be overriden by BoundaryProcess types to indicate the type of boundary condition, e.g:

BoundaryStyle(::Type{BP}) = Dirichlet()

where BP is a BoundaryProcess that provides the boundary conditions.

CryoGrid.Interface.Coupled — Type

Coupled{P1,P2} = System{Tuple{T1,T2}} where {T1,T2}

Represents a coupled pair of explicitly processes. Alias for System{Tuple{P1,P2}}. Coupled provides a simple mechanism for defining new behaviors on composite processes/systems.

CryoGrid.Interface.Coupled — Method

Coupled(p1,p2)

Alias for System(p1,p2).

CryoGrid.Interface.Dirichlet — Type

BoundaryStyle instance for Dirichlet boundary conditions.

CryoGrid.Interface.Layer — Type

Abstract base type for all layers.

CryoGrid.Interface.Neumann — Type

BoundaryStyle instance for Neumann boundary conditions.

CryoGrid.Interface.Process — Type

Abstract base type for all dynamical processes.

CryoGrid.Interface.SubSurface — Type

SubSurface <: Layer

Abstract base type for layers in the stratigraphy, e.g. soil, snow, pond, etc.

CryoGrid.Interface.SubSurfaceProcess — Type

SubSurfaceProcess <: Process

Abstract base type for subsurface processes, i.e. processes that operate at or below the surface, such as heat conduction, water infiltration, etc.

CryoGrid.Interface.System — Type

System{TProcs} <: Process

Represents a explicitly or implicitly coupled system of processes. TProcs is always a Tuple of other processes.

CryoGrid.Interface.Top — Type

Top <: Layer

Generic "top" layer that marks the upper boundary of the subsurface grid.

CryoGrid.Interface.Boundary — Type

Boundary = Union{Top,Bottom}

Alias that refers to the type union over both Top and Bottom layer types.

CryoGrid.Interface.diagnosticstep! — Method

diagnosticstep!(l::Layer, p::Process, state)

Defines the diagnostic update for a Process on a given Layer.

CryoGrid.Interface.initialcondition! — Method

initialcondition!(::Layer, state)

Defines the initial condition for a given Layer. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.Interface.initialcondition! — Method

initialcondition!(::Layer, ::Process, state)

Defines the initial condition for a given Process on Layer. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.Interface.initialcondition! — Method

initialcondition!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines the initial condition for two Processes on adjacent layers. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.Interface.interact! — Method

interact!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines a boundary interaction between two processes on adjacent layers.

CryoGrid.Interface.prognosticstep! — Method

prognosticstep!(l::Layer, p::Process, state)

Defines the prognostic update for a Process on a given layer. Note that an instance of prognosticstep! must be provided for all non-boundary (subsurface) processes/layers.

CryoGrid.Interface.variables — Method

variables(::Layer, ::Process)

Defines variables for a given Process on Layer. Implementations should return a Tuple of CryoGrid.Common.Vars.

CryoGrid.Interface.variables — Method

variables(::Layer)

Defines variables for a given Layer type. Implementations should return a Tuple of CryoGrid.Common.Vars.