Hydrology

CryoGrid.Hydrology.BucketScheme — Type

BucketScheme <: WaterFlow

"Bucket" water scheme for downward advective flow due to gravity.

CryoGrid.Hydrology.DampedET — Type

DampedET{Tftr,Tfev,Tdtr,Tdev}

Corresponds to evapotranspiration scheme 2 described in section 2.2.4 of Westermann et al. (2022).

CryoGrid.Hydrology.Evapotranspiration — Type

Evapotranspiration

Base type for parameterizations of evapotranspiration (ET).

CryoGrid.Hydrology.HydraulicProperties — Type

HydraulicProperties

Default material hydraulic properties.

CryoGrid.Hydrology.NoFlow — Type

NoFlow <: WaterFlow

Represents a zero flow scheme

CryoGrid.Hydrology.WaterBalance — Type

WaterBalance{TFlow<:WaterFlow,TET<:Union{Nothing,Evapotranspiration},Tdt,Taux,TProp} <: CryoGrid.SubSurfaceProcess

Represents subsurface water transport processes.

CryoGrid.Hydrology.WaterBalanceProperties — Type

WaterBalanceProperties

Numerical constants shared across water balance implementations.

CryoGrid.Hydrology.WaterFlow — Type

WaterFlow

Base type for different formulations of water flow in WaterBalance.

CryoGrid.Hydrology.WaterTableInitializer — Type

WaterTableInitializer <: VarInitializer{:sat}

Simple, piecewise constant initializer for saturation state that takes a surface-level saturation sat0 and water table depth z_tab and produces a two-segment, piecewise constant profile with the saturation level set to (sat0 + 1.0) / 2 from the halfway point down to the water table.

CryoGrid.Hydrology.hydraulicconductivity! — Method

hydraulicconductivity!(sub::SubSurface, water::WaterBalance, state)

Computes hydraulic conductivities for the given subsurface layer and water balance scheme.

CryoGrid.Hydrology.hydraulicproperties — Method

hydraulicproperties(::SubSurface)

Retrieves the hydraulic properties from the given subsurface layer. Default implementation simply returns the default configuration of HydraulicProperties.