Surface Energy Balance

CryoGrid.Surface.SurfaceEnergyBalance — Type

SurfaceEnergyBalance{TSolution,TStabFun,TPara,F} <: BoundaryProcess{HeatBalance}

Surface energy balance upper boundary condition.

CryoGrid.Surface.SurfaceWaterBalance — Type

SurfaceWaterBalance{TR,TS} <: BoundaryProcess{Union{WaterBalance, SnowMassBalance}}

The SurfaceWaterBalance represents the closure of the water balance at the surface and acts as both a Neumann-type upper boundary condition for snow and water fluxes as well as an accountant for the water mass balance at the surface.