Snow

CryoGrid.Snow.DynamicSnowMassBalance — Type

DynamicSnowMassBalance{TAcc,TAbl} <: SnowMassBalance

Dynamic snow mass balance, i.e. where snow is accumulated and ablated according to dynamic physical processes.

CryoGrid.Snow.PrescribedSnowMassBalance — Type

PrescribedSnowMassBalance{Tswe} <: SnowMassBalance

"Prescribed" snow mass balance, i.e. where the snow water equivalent is given as a constant or forcing.

CryoGrid.Snow.SnowMassBalance — Type

SnowMassBalance{Tpara<:SnowMassParameterization} <: CryoGrid.SubSurfaceProcess

Base type for subsurface processes representing the dynamic accumulation and ablation of snow cover.

CryoGrid.Snow.Snowpack — Type

Snowpack{Tpara<:SnowpackParameterization,Tmass<:SnowMassBalance,Theat<:HeatBalance,Twater<:WaterBalance,Taux} <: CryoGrid.SubSurface

Generic representation of a snowpack "subsurface" layer.

CryoGrid.Snow.snowdensity — Method

snowdensity(::Snowpack, state)

Retrieve the current snow density.

CryoGrid.Snow.swe — Method

swe(::Snowpack, ::SnowMassBalance, state)

Retrieve the current snow water equivalent of the snowpack.