Soils

CryoGrid.Soils.AbstractGround — Type

AbstractGround{Tpara<:GroundParameterization,Theat<:Optional{HeatBalance},Twater<:Optional{WaterBalance}} <: SubSurface

Base type for all ground layers defining heat and water balance schemes.

CryoGrid.Soils.Ground — Type

Ground{Tpara,Theat<:Optional{HeatBalance},Twater<:Optional{WaterBalance},Taux} <: Soil{Tpara,Theat,Twater}

Generic representation of a Ground layer with material parameterization para.

CryoGrid.Soils.Ground — Method

Ground(soilprofile::SoilProfile; kwargs...)

CryoGrid.Soils.Heterogeneous — Type

Heterogeneous{V,N,D,Taux} <: SoilParameterization

Special SoilParameterization which wraps a Profile of another soil parameterization type to indicate that it should be heterogeneous with over depth.

CryoGrid.Soils.MineralOrganic — Type

MineralOrganic{Tpor,Tsat,Torg} <: SoilParameterization

Represents a simple organic/mineral soil mixutre in terms of its characteristic fractions: i.e. natural porosity, saturation, and organic solid fraction. This is the standard CryoGrid representation of a discrete soil volume.

CryoGrid.Soils.RichardsEq — Type

RichardsEq{Tform<:RichardsEqFormulation,Tswrc<:SWRC,Taux,TΩ} <: Hydrology.WaterFlow

The Richardson-Richards equation describes the flow of water in porous media under unsaturated condition.

CryoGrid.Soils.Soil — Type

Soil{Tpara,Theat,Twater}

Type alias for any AbstractGround layer with parameterization of type SoilParameterization.

CryoGrid.Soils.SoilParameterization — Type

SoilParameterization

Base type for parameterizations of soil consituents.

CryoGrid.Soils.SoilProfile — Type

SoilProfile{N,IT,VT} = Profile{N,IT,VT} where {N,IT<:NTuple{N},VT<:NTuple{N,SoilParameterization}}

Alias for depthwise Profile where the values are SoilParameterization types.

CryoGrid.Soils.SoilProfile — Method

SoilProfile(pairs::Pair{<:DistQuantity,<:SoilParameterization}...)

Alias for Profile(pairs...) specific for SoilProfiles.

CryoGrid.Soils.SoilTexture — Type

SoilTexture{Tsand,Tsilt,Tclay}

Represents soil "texture" as a simple mixture of sand, silt, and clay.

CryoGrid.Soils.mineral — Method

mineral(soil::Soil, state, i)
mineral(soil::Soil, state)
mineral(soil::Soil)

Retrieves the volumetric mineral content for the given layer at grid cell i, if provided.

CryoGrid.Soils.organic — Method

organic(soil::Soil, state, i)
organic(soil::Soil, state)
organic(soil::Soil)

Retrieves the volumetric organic content for the given layer at grid cell i, if provided.

CryoGrid.Soils.porosity — Method

porosity(soil::Soil, state, i)
porosity(soil::Soil, state)
porosity(soil::Soil)

Retrieves the porosity for the given layer at grid cell i, if provided.