Soils

CryoGrid.Physics.Soils.CharacteristicFractions — Type

CharacteristicFractions{P1,P2,P3,P4} <: SoilParameterization

Represents uniform composition of a soil volume in terms of fractions: excess ice, natural porosity, saturation, and organic solid fraction.

CryoGrid.Physics.Soils.Soil — Type

Basic Soil layer.

CryoGrid.Physics.Soils.SoilParameterization — Type

SoilParameterization

Abstract base type for parameterizations of soil properties.

CryoGrid.Physics.Soils.mineral — Method

mineral(soil::Soil, state, i)

Retrieves the mineral content for the given layer at grid cell i, if provided. Defaults to using the scalar mineral content defined on soil.

CryoGrid.Physics.Soils.organic — Method

organic(soil::Soil, state, i)

Retrieves the organic content for the given layer at grid cell i, if provided. Defaults to using the scalar organic content defined on soil.

CryoGrid.Physics.Soils.porosity — Method

porosity(soil::Soil, state, i)

Retrieves the porosity for the given layer at grid cell i, if provided. Defaults to using the scalar porosity defined on soil.