Water balance

CryoGrid.Physics.WaterBalance.SWRC — Type

SWRC{F,∇F}

Soil water retention curve with function type F and derivative ∇F.

CryoGrid.Physics.WaterBalance.SWRC — Method

SWRC(f::SWRCFunction)

Convenience constructor for SWRC that automatically generates an analytical derivative of the given retention curve function f using ModelingToolkit/Symbolics.jl.

CryoGrid.Physics.WaterBalance.VanGenuchten — Type

VanGenuchten <: SWRCFunction

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.