InputOutput

This module provides types and utilities for reading input files (i.e. parameters and forcings) as well as saving model outputs.

CryoGrid.InputOutput.CryoGridOutput — Type

CryoGridOutput{TSol}

Helper type that stores the raw output from a CryoGrid run along with DimArray views of all logged variables. CryoGridOutput overrides Base.getproperty to allow for direct dot-syntax access of state variables. For example, if your model has a grid variable named T, out.T returns a DimArray with indexed time and depth axes. For OrdinaryDiffEq.jl outputs, the ODESolution can be accessed via out.sol, or for convenience, the continuous solution at time t can be computed via out(t) which is equivalent to withaxes(out.sol(t)).

CryoGrid.InputOutput.CryoGridOutput — Method

Evaluates the continuous solution at time t.

CryoGrid.InputOutput.CryoGridParams — Type

CryoGridParams{T,TM} <: DenseArray{T,1}

Wraps a ModelParameters.Model parameter handler for CryoGrid types. It is recommended not to use this type directly in math or linear algebra operations but rather to use Base.values to obtain a normal array of parameter values.

CryoGrid.InputOutput.CryoGridParams — Method

CryoGridParams(obj)

Constructs a modelParameters.Model wrapped with CryoGridParams from obj. If full_metadata is true, additonal fields for nested Parameterization types will be added.

CryoGrid.InputOutput.FixedParam — Type

FixedParam(p::NamedTuple)
FixedParam(; kw...)
FixedParam(val)

Subtype of AbstractParam that rerpesents a "fixed" parameter which should not be included in the set of free parameters for the model.

CryoGrid.InputOutput.ForcingFormat — Type

Represents an externally specified format for forcing inputs. IO functions should dispatch on specific types T<:ForcingFormat that they implement.

CryoGrid.InputOutput.ForcingFormatJSON — Type

JSON forcing input format (from CryoGridLite) with specified version indicator.

CryoGrid.InputOutput.ForcingFormatNCD — Type

NetCDF forcing input format.

CryoGrid.InputOutput.Input — Type

Input{name,attrsType<:NamedTuple}

The Input type represents a placeholder for (typically time-varying) inputs that will be replaced with the numeric value at runtime. The name of the Input will be used to match it to the corresponding input function.

CryoGrid.InputOutput.InputFunctionProvider — Type

InputFunctionProvider

Generic input provider that simply wraps a NamedTuple of functions.

CryoGrid.InputOutput.InputProvider — Type

InputProvider

Base type for Input providers that realize one or more Inputs.

CryoGrid.InputOutput.LinearTrend — Type

LinearTrend{id,TSlope,TInt} <: DynamicParameterization

Applies a linear trend to a parameter p by reparameterizing it as: p = p₁*t + p₀

CryoGrid.InputOutput.ParamsJSON — Type

JSON parameter input format (from CryoGridLite). Not yet implemented.

CryoGrid.InputOutput.ParamsYAML — Type

YAML parameter input format matching that of the CryoGrid community model. Not yet implemented.

CryoGrid.InputOutput.PiecewiseLinear — Type

PiecewiseLinear{id,N,T0,TKnots,I} <: DynamicParameterization

Reparameterizes parameter p as p = p₁δ₁t + ⋯ + pₖδₖt where δₖ are indicators for when tₖ₋₁ <= t <= tₖ. To facilitate sampling and optimization, change points tᵢ are parameterized as bin widths, which should be strictly positive. PiecewiseLinear will normalize them and scale by the size of the time interval.

CryoGrid.InputOutput.Transformed — Type

Transformed{F,NT} <: DynamicParameterization

Reparameterization corresponding to an arbitrary functional transformation.

CryoGrid.InputOutput.PiecewiseConstant — Method

Special case of PiecewiseLinear.

CryoGrid.InputOutput.inputs — Method

inputs(; named_inputs...)

Alias for the constructor of InputFunctionProvider.

CryoGrid.InputOutput.loadforcings — Method

loadforcings(filename::String)
loadforcings(resource::Resource; outdir=DEFAULT_FORCINGS_DIR)
loadforcings([format::ForcingFormat], filename::String; outdir=DEFAULT_FORCINGS_DIR)

Loads forcing data from the given file according to the format specified by format. By default, the forcing format is automatically detected via detectformat. Returns a DimStack containing all forcing data and metadata