Numerics

CryoGrid.Numerics.Algebraic — Type

Algebraic{name,S,T,units,domain} <: Var{name,S,T,units,domain}

Defines an algebraic (implicit) state variable.

CryoGrid.Numerics.Delta — Type

Delta{dname,name,S,T,units,domain} <: Var{dname,S,T,units,domain}

Defines a "delta" term du for variable u, which is the time-derivative/divergence for prognostic variables or the residual for algebraic variables.

CryoGrid.Numerics.Diagnostic — Type

Diagnostic{name,S,T,units,domain} <: Var{name,S,T,units,domain}

Defines a diagnostic variable which is allocated and cached per timestep but not integrated over time.

CryoGrid.Numerics.DiffCache — Type

DiffCache{N,A,Adual}

Wrapper around PreallocationTools.DiffCache that stores state variables in forward-diff compatible cache arrays.

CryoGrid.Numerics.FunctionInitializer — Type

FunctionInitializer{varname,F} <: VarInitializer{varname}

Initializes a scalar or vector-valued state variable using an arbitrary method/function.

CryoGrid.Numerics.Grid — Type

struct Grid{S,G,Q,A} <: AbstractDiscretization{Q,1}

Represents the 1D spatial discretization on which time integration is performed. S is a GridSpec, either Edges or Cells (always edges upon initial construction). The grid representation can be converted (allocation free) between grid edges and cells via the cells and edges methods. G represents the geometry/volume on which the vertical 1D discretization is applied. A is the underlying array type, and Q is the numerical type (e.g. Float64 or a Unitful.Quantity).

CryoGrid.Numerics.InterpInitializer — Type

InterpInitializer{varname,P,I,E} <: VarInitializer{varname}

Initializer for on-grid variables that takes a Profile as initial values and interpolates along the model grid. The interpolation mode is linear by default, but can also be any other Gridded interpolation mode supported by Interpolations.jl.

CryoGrid.Numerics.Prognostic — Type

Prognostic{name,S,T,units,domain} <: Var{name,S,T,units,domain}

Defines a prognostic (time-integrated) state variable.

CryoGrid.Numerics.VarStates — Type

VarStates{names,griddvars,TU,TD,TV,DF,DG}

Generic container for holding discretized state arrays for declared variables (Var types), as well as the prototype prognostic state vector (uproto).

CryoGrid.Numerics.VarStates — Method

VarStates(vars::GroupedVars, D::Numerics.AbstractDiscretization, chunksize::Int, arrayproto::Type{A}=Vector) where {A<:AbstractVector}

CryoGrid.Numerics.discretize — Method

discretize([::Type{A}], ::T,  ::Var) where {T,N,D<:AbstractDiscretization{T,N},A<:AbstractArray{T,N}}

Produces a discretization of the given variable based on T and array type A.

CryoGrid.Numerics.divergence! — Method

divergence!(dx::AbstractVector, j::AbstractVector, Δj::AbstractVector)

Calculates the first-order divergence over a 1D flux vector field j and grid cell lengths Δj. Divergences are added to existing values in dx.

CryoGrid.Numerics.dual — Method

dual(x::Number, ::Type{tag}) where {tag}
dual(x::A, ::Type{tag}) where {N,T,A<:SVector{N,T},tag}

Constructs a ForwardDiff.Dual number (or static array thereof) with tag from x.

CryoGrid.Numerics.flux! — Method

flux!(j::AbstractVector, x::AbstractVector, Δx::AbstractVector, k::AbstractVector)

Calculates the first-order, non-linear spatial flux over a discretized variable x with conductivity k. x is assumed to have shape (N,), Δx shape (N-1,), and j and k shape (N+1,) such that j[2:end-1] represents the fluxes over the inner grid cell faces. Fluxes are added to existing values in j.

CryoGrid.Numerics.harmonicmean! — Method

harmonicmean!(h::AbstractVector, x::AbstractVector, w::AbstractVector)

Vectorized harmonic mean of elements in x with weights w. Output is stored in h, which should have size length(x)-1.

CryoGrid.Numerics.harmonicmean — Method

harmonicmean(x₁, x₂, w₁, w₂)

Simple weighted harmonic mean of two values, x₁ and x₂.

CryoGrid.Numerics.heaviside — Method

heaviside(x)

Differentiable implementation of heaviside step function, i.e:

$h(x) = \begin{cases} 1 & x ≥ 0 \\ 0 & x < 0 \end{cases}$

CryoGrid.Numerics.initializer — Method

initializer(varname::Symbol, x::Number) => FunctionInitializer w/ constant
initializer(varname::Symbol, f::Function) => FunctionInitializer
initializer(varname::Symbol, profile::Profile, interp=Linear(), extrap=Flat()) => InterpInitializer

Convenience constructor for VarInitializer that selects the appropriate initializer type based on the arguments.

CryoGrid.Numerics.logistic — Method

logistic(x)

Numerically stable logistic function.

$σ(x) = \begin{cases} \frac{1}{1+\exp(-x)} & x ≥ 0 \\ \frac{\exp(x)}{1+\exp(x)} & x < 0 \end{cases}$

CryoGrid.Numerics.logit — Method

logit(x)

Numerically stable logit function. True domain is (0,1) but inputs are clamped to (ϵ,1-ϵ) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.nonlineardiffusion! — Method

nonlineardiffusion!(dx::AbstractVector, j::AbstractVector, x::AbstractVector, Δx::AbstractVector, k::AbstractVector, Δk::AbstractVector)

Fast alternative to flux! and divergence! which computes fluxes and divergences (via _flux_kernel and _div_kernel) in a single pass. Note, however, that loop vectorization with @turbo is not possible because of necessary loop-carried dependencies. Fluxes and divergences are added to the existing values stored in j and dx.

CryoGrid.Numerics.softplus — Method

softplus(x)

Numerically stable softplus function.

$s(x) = \log(1+\exp(-|x|)) + \max(x,ϵ)$

CryoGrid.Numerics.softplusinv — Method

softplusinv(x)

Numerically stable softplus inverse function. True domain is (0,∞) but inputs are clamped to (ϵ,∞) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.tdma_solve! — Function

tdma_solve!(x, a, b, c, d)

Tridiagonal matrix solver; borrowed from CryoGridLite. Modifies all input vectors in-place.

CryoGrid.Numerics.updategrid! — Method

updategrid!(grid::Grid{Edges,G,Q}, vals::Q=grid) where {G,Q}

Overwrites grid edges with vals, and recomputes grid centers/deltas to be consistent with the new grid.

CryoGrid.Numerics.∇ — Method

∇(f::F, x::AbstractArray) where {F}

Takes a function y = f(x) and vector-valued argument x and returns a tuple: (y, ∇ₓy). The gradient is calculated using forward-mode automatic differentiation.

CryoGrid.Numerics.∇ — Method

∇(f::F, x::Number) where {F}

Takes a function y = f(x) and argument x and returns a tuple: (y, ∂y∂x). The derivative is calculated using forward-mode automatic differentiation.