Numerics

CryoGrid.Numerics.AutoGrid — Type

AutoGrid <: DiscretizationStrategy

CryoGrid.Numerics.DiffCache — Type

DiffCache{TCache}

Wrapper around PreallocationTools.DiffCache that stores state variables in forward-diff compatible cache arrays.

CryoGrid.Numerics.DiscretizationStrategy — Type

Base type for discretization "strategies" that generate a spatial discretization (typically a grid).

CryoGrid.Numerics.Grid — Type

struct Grid{S,G,Q,A} <: AbstractDiscretization{Q,1}

Represents the 1D spatial discretization on which time integration is performed. S is a GridOffset, either Edges or Cells (always edges upon initial construction). The grid representation can be converted (allocation free) between grid edges and cells via the cells and edges methods. G represents the geometry/volume on which the vertical 1D discretization is applied. A is the underlying array type, and Q is the numerical type (e.g. Float64 or a Unitful.Quantity).

CryoGrid.Numerics.PresetGrid — Type

Simple discretization strategy that just supplies a pre-specified grid.

CryoGrid.Numerics.StateVars — Type

StateVars{names,griddvars,TU,TD,TV,DF,DG}

Generic container for holding discretized state arrays for declared variables (Var types), as well as the prototype prognostic state vector (uproto).

CryoGrid.Numerics.StateVars — Method

StateVars(vars::GroupedVars, D::Numerics.AbstractDiscretization, chunk_size::Int, arrayproto::Type{A}=Vector) where {A<:AbstractVector}

CryoGrid.Numerics.build_mass_matrix — Method

build_mass_matrix(u::ComponentVector, states::StateVars)

Constructs a mass matrix M⋅∂u∂t = f(u) suitable for the prognostic state vector u based on the defined variable types.

CryoGrid.Numerics.divergence! — Method

divergence!(dx::AbstractVector, j::AbstractVector, Δj::AbstractVector)

Calculates the first-order divergence over a 1D flux vector field j and grid cell lengths Δj. Divergences are added to existing values in dx.

CryoGrid.Numerics.dual — Method

dual(x::Number, ::Type{tag}) where {tag}
dual(x::A, ::Type{tag}) where {N,T,A<:SVector{N,T},tag}

Constructs a ForwardDiff.Dual number (or static array thereof) with tag from x.

CryoGrid.Numerics.flux! — Method

flux!(j::AbstractVector, x::AbstractVector, Δx::AbstractVector, k::AbstractVector)

Calculates the first-order, non-linear spatial flux over a discretized variable x with conductivity k. x is assumed to have shape (N,), Δx shape (N-1,), and j and k shape (N+1,) such that j[2:end-1] represents the fluxes over the inner grid cell faces. Fluxes are added to existing values in j.

CryoGrid.Numerics.harmonicmean! — Method

harmonicmean!(h::AbstractVector, x::AbstractVector, w::AbstractVector)

Vectorized harmonic mean of elements in x with weights w. Output is stored in h, which should have size length(x)-1.

CryoGrid.Numerics.harmonicmean — Method

harmonicmean(x₁, x₂, w₁, w₂)

Simple weighted harmonic mean of two values, x₁ and x₂.

CryoGrid.Numerics.heaviside — Method

heaviside(x)

Differentiable implementation of heaviside step function, i.e:

$h(x) = \begin{cases} 1 & x ≥ 0 \\ 0 & x < 0 \end{cases}$

CryoGrid.Numerics.instantiate — Method

instantiate(::Var, ::T, ::Type{A}) where {T,N,D<:AbstractDiscretization{T,N},A<:AbstractArray{T,N}}

Produces a discretization of the given variable based on T and array type A.

CryoGrid.Numerics.logistic — Method

logistic(x)

Numerically stable logistic function.

$σ(x) = \begin{cases} \frac{1}{1+\exp(-x)} & x ≥ 0 \\ \frac{\exp(x)}{1+\exp(x)} & x < 0 \end{cases}$

CryoGrid.Numerics.logit — Method

logit(x)

Numerically stable logit function. True domain is (0,1) but inputs are clamped to (ϵ,1-ϵ) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.makegrid — Method

makegrid(strategy::DiscretizationStrategy, bounds::NTuple{2,<:DistQuantity})
makegrid(layer::Layer, strategy::DiscretizationStrategy, bounds::NTuple{2,<:DistQuantity})

Constructs a Grid spanning bounds using the given strategy. makegrid can also be specified for specific Layer types when layers have specific discretization requirements.

CryoGrid.Numerics.nonlineardiffusion! — Method

nonlineardiffusion!(dx::AbstractVector, j::AbstractVector, x::AbstractVector, Δx::AbstractVector, k::AbstractVector, Δk::AbstractVector)

Fast alternative to flux! and divergence! which computes fluxes and divergences (via _flux_kernel and _div_kernel) in a single pass. Note, however, that loop vectorization with @turbo is not possible because of necessary loop-carried dependencies. Fluxes and divergences are added to the existing values stored in j and dx.

CryoGrid.Numerics.softplus — Method

softplus(x)

Numerically stable softplus function.

$s(x) = \log(1+\exp(-|x|)) + \max(x,ϵ)$

CryoGrid.Numerics.softplusinv — Method

softplusinv(x)

Numerically stable softplus inverse function. True domain is (0,∞) but inputs are clamped to (ϵ,∞) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.tdma_solve! — Function

tdma_solve!(x, a, b, c, d)

Tridiagonal matrix solver via the Thomas algorithm; adapted from original CryoGridLite implementation. Modifies all input vectors in-place.

CryoGrid.Numerics.updategrid! — Method

updategrid!(grid::Grid{Edges,G,Q}, vals::Q=grid) where {G,Q}

Overwrites grid edges with vals, and recomputes grid centers/deltas to be consistent with the new grid.

CryoGrid.Numerics.∇ — Method

∇(f::F, x::AbstractArray) where {F}

Takes a function y = f(x) and vector-valued argument x and returns a tuple: (y, ∇ₓy). The gradient is calculated using forward-mode automatic differentiation.

CryoGrid.Numerics.∇ — Method

∇(f::F, x::Number) where {F}

Takes a function y = f(x) and argument x and returns a tuple: (y, ∂y∂x). The derivative is calculated using forward-mode automatic differentiation.