Numerics

CryoGrid.Numerics.ConstantInit — Type

ConstantInit{varname,T} <: VarInit

Initializes a scalar or vector-valued state variable with a pre-specified constant value.

CryoGrid.Numerics.DiffCache — Type

DiffCache{N,A,Adual}

Extension of PreallocationTools.DiffCache that stores state variables in forward-diff compatible cache arrays.

CryoGrid.Numerics.Grid — Type

struct Grid{S,G,Q,A} <: AbstractDiscretization{Q,1}

Represents the 1D spatial discretization on which time integration is performed. S is a GridSpec, either Edges or Cells (always edges upon initial construction). The grid representation can be converted (allocation free) between grid edges and cells via the cells and edges methods. G represents the geometry/volume on which the vertical 1D discretization is applied. A is the underlying array type, and Q is the numerical type (e.g. Float64 or a Unitful.Quantity).

CryoGrid.Numerics.InterpInit — Type

InterpInit{varname,P,I,E} <: VarInit

Init for on-grid variables that takes a Profile as initial values and interpolates along the model grid. The interpolation mode is linear by default, but can also be quadratic, cubic, or any other Gridded interpolation mode supported by Interpolations.jl.

CryoGrid.Numerics.VarStates — Type

VarStates{names,griddvars,TU,TD,TV,DF,DG}

Generic container for holding discretized state arrays for declared variables (Var types), as well as the prototype prognostic state vector (uproto).

CryoGrid.Numerics.VarStates — Method

VarStates(vars::GroupedVars, D::Numerics.AbstractDiscretization, chunksize::Int, arrayproto::Type{A}=Vector) where {A<:AbstractVector}

CryoGrid.Numerics.Tabulated — Method

Tabulated(f, argknots...)

Alias for tabulate intended for function types.

CryoGrid.Numerics.discretize — Method

discretize([::Type{A}], ::T,  ::Var) where {T,N,D<:AbstractDiscretization{T,N},A<:AbstractArray{T,N}}

Produces a discretization of the given variable based on T and array type A.

CryoGrid.Numerics.finitediff! — Method

finitediff!(∂x::AbstractVector, x::AbstractVector, Δ::AbstractVector)

First order forward finite difference operator.

CryoGrid.Numerics.harmonicmean! — Method

harmonicmean!(h::AbstractVector, x::AbstractVector, w::AbstractVector)

Vectorized harmonic mean of elements in x with weights w. Output is stored in h, which should have size length(x)-1.

CryoGrid.Numerics.harmonicmean — Method

harmonicmean(x₁, x₂, w₁, w₂)

Simple weighted harmonic mean of two values, x₁ and x₂.

CryoGrid.Numerics.heaviside — Method

heaviside(x)

Differentiable implementation of heaviside step function, i.e:

$h(x) = \begin{cases} 1 & x ≥ 0 \\ 0 & x < 0 \end{cases}$

CryoGrid.Numerics.init! — Method

init!(x::AbstractVector, init::VarInit, args...)

Initializes state variable x using initializer init and implementation-specific additional arguments.

CryoGrid.Numerics.initializer — Method

initializer(varname::Symbol, value::T) where {T} => ConstantInit
initializer(varname::Symbol, profile::Profile, interp=Linear(), extrap=Flat()) => InterpInit

Convenience constructor for VarInits that selects the appropriate initializer type based on the arguments.

CryoGrid.Numerics.lineardiffusion! — Method

lineardiffusion!(∂y::AbstractVector, x::AbstractVector, Δ::AbstractVector, k::Number)

Second order Laplacian with constant diffusion k.

CryoGrid.Numerics.logistic — Method

logistic(x)

Numerically stable logistic function.

$σ(x) = \begin{cases} \frac{1}{1+\exp(-x)} & x ≥ 0 \\ \frac{\exp(x)}{1+\exp(x)} & x < 0 \end{cases}$

CryoGrid.Numerics.logit — Method

logit(x)

Numerically stable logit function. True domain is (0,1) but inputs are clamped to (ϵ,1-ϵ) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.nonlineardiffusion! — Method

nonlineardiffusion!(
    ∂y::AbstractVector{Float64},
    x::AbstractVector{Float64}, 
    Δx::AbstractVector{Float64},
    k::AbstractVector{Float64},
    Δk::AbstractVector{Float64}
)

Second order Laplacian with non-linear diffusion operator, k. Accelerated using LoopVectorization.@turbo for Float64 vectors.

CryoGrid.Numerics.profile2array — Method

profile2array(profile::Profile{N,D,T};names) where {N,D,T}

Constructs a DimArray from the given Profile, i.e. pairs Q => (x1,...,xn) where x1...xn are the values defined at Q. Column names for the resulting DimArray can be set via the names parameter which accepts an NTuple of symbols, where N must match the number of parameters given (i.e. n).

CryoGrid.Numerics.softplus — Method

softplus(x)

Numerically stable softplus function.

$s(x) = \log(1+\exp(-|x|)) + \max(x,ϵ)$

CryoGrid.Numerics.softplusinv — Method

softplusinv(x)

Numerically stable softplus inverse function. True domain is (0,∞) but inputs are clamped to (ϵ,∞) for numerical convenience, making the effective domain (-∞,∞).

CryoGrid.Numerics.tabulate — Method

tabulate(f, argknots::Pair{Symbol,<:Union{Number,AbstractArray}}...)

Tabulates the given function f using a linear, multi-dimensional interpolant. Knots should be given as pairs :arg => A where A is a StepRange or Vector of input values (knots) at which to evaluate the function. A may also be a Number, in which case a pseudo-point interpolant will be used (i.e valid on [A,A+ϵ]). No extrapolation is provided by default but can be configured via Interpolations.extrapolate.

CryoGrid.Numerics.∇ — Method

∇(f, dvar::Symbol)

Automatically generates an analytical partial derivative of f w.r.t dvar using Symbolics.jl. To avoid symbolic tracing issues, the function should 1) be pure (no side effects or non-mathematical behavior) and 2) avoid indeterminate control flow such as if-else or while blocks (technically should work but sometimes doesn't...). Conditional logic can be included using IfElse.ifelse. Additional argument names are extracted automatically from the method signature of f. Keyword arg choosefn should be a function which selects from available methods of f (returned by methods); defaults to first. Note that ∇ uses RuntimeGeneratedFunction to produce a fully specialized and compiled Julia function; it may be slow on the first call (due to compilation), but should be just as fast as handwriting it on subsequent calls.

Example:

For $f(x,y) = 2x + xy$, $\frac{\partial f}{\partial x} = 2 + y$. Using ∇, we can obtain this automagically:

f(x,y) = 2*x + x*y
∇f_x = ∇(f,:x)
∇f_x(2.0,3.0)

# output

5.0