Snow

CryoGrid.Snow.SnowBC — Type

SnowBC

Type alias for any BoundaryProcess compatible with SnowMassBalance.

CryoGrid.Snow.SnowMassBalance — Type

SnowMassBalance{TAcc,TAbl} <: CryoGrid.SubSurfaceProcess

Subsurface process for snow layers governing how snow is accumulated and ablated.

CryoGrid.Snow.Snowpack — Type

Snowpack{Tpara<:SnowpackParameterization,Tmass<:SnowMassBalance,Twater<:WaterBalance,Theat<:HeatBalance,Taux} <: CryoGrid.SubSurface

Generic representation of a snowpack "subsurface" layer.

CryoGrid.Snow.Snowpack — Method

Snowpack(para::SnowpackParameterization; kwargs...)

Convenience constructor that accepts the parameterization as a positional argument.

CryoGrid.Snow.ablation! — Method

ablation!(::Top, ::SnowBC, ::Snowpack, ::SnowMassBalance, stop, ssnow)

Computes snow mass balance fluxes due to ablation (e.g. snow melt).

CryoGrid.Snow.accumulation! — Method

accumulation!(::Top, ::SnowBC, ::Snowpack, ::SnowMassBalance, stop, ssnow)

Computes snow mass balance fluxes due to accumulation (e.g. snowfall).

CryoGrid.Snow.snowdensity — Method

snowdensity(::Snowpack, state)

Retrieve the current snow density.

CryoGrid.Snow.snowdepth — Method

snowdepth(::Snowpack, ::SnowMassBalance, state)

Retrieve the current snow depth.

CryoGrid.Snow.snowfall — Method

snowfall(::SnowBC, state)

Retrieves the current snowfall flux for the given SnowBC. Defaults to state.jw_snow.