Soils

CryoGrid.Physics.Soils.DallAmico — Type

DallAmico <: SFCCFunction

Dall'Amico M, 2010. Coupled water and heat transfer in permafrost modeling. Ph.D. Thesis, University of Trento, pp. 43.

CryoGrid.Physics.Soils.McKenzie — Type

McKenzie <: SFCCFunction

McKenzie JM, Voss CI, Siegel DI, 2007. Groundwater flow with energy transport and water-ice phase change: numerical simulations, benchmarks, and application to freezing in peat bogs. Advances in Water Resources, 30(4): 966–983. DOI: 10.1016/j.advwatres.2006.08.008.

CryoGrid.Physics.Soils.SFCC — Type

SFCC{F,∇F,S} <: FreezeCurve

Generic representation of the soil freeze characteristic curve. The shape and parameters of the curve are determined by the implementation of SFCCFunction f. Also requires an implementation of SFCCSolver which provides the solution to the non-linear mapping H <–> T.

CryoGrid.Physics.Soils.SFCC — Type

SFCC(f::SFCCFunction, s::SFCCSolver=SFCCNewtonSolver())

Convenience constructor for SFCC that automatically generates an analytical derivative of the given freeze curve function f using ModelingToolkit/Symbolics.jl. To avoid symbolic tracing issues, the function should 1) be pure (no side effects or non-mathematical behavior) and 2) avoid indeterminate control flow such as if-else or while blocks (technically should work but sometimes doesn't...). Conditional logic can be incorporated via IfElse.ifelse. See the documentation for Symbolics.jl for more information and technical details.

CryoGrid.Physics.Soils.SFCC — Type

SFCC(f::SFCCTable, s::SFCCSolver=SFCCNewtonSolver())

Constructs a SFCC from the precomputed SFCCTable. The derivative is generated using the gradient function provided by Interpolations.

CryoGrid.Physics.Soils.SFCC — Method

Updates state variables according to the specified SFCC function and solver. For heat conduction with enthalpy, this is implemented as a simple passthrough to the non-linear solver. For heat conduction with temperature, we can simply evaluate the freeze curve to get C_eff, θl, and H.

CryoGrid.Physics.Soils.SFCCNewtonSolver — Type

Specialized implementation of Newton's method with backtracking line search for resolving the energy conservation law, H = TC + Lθ. Attempts to find the root of the corresponding temperature residual: ϵ = T - (H - Lθ(T)) / C(θ(T)) and uses backtracking to avoid jumping over the solution. This prevents convergence issues that arise due to discontinuities and non-monotonic behavior in most common soil freeze curves.

CryoGrid.Physics.Soils.Soil — Type

Basic Soil layer.

CryoGrid.Physics.Soils.SoilCharacteristicFractions — Type

SoilCharacteristicFractions{P1,P2,P3,P4} <: SoilParameterization

Represents the composition of the soil in terms of fractions: excess ice, natural porosity, saturation, and organic/(mineral + organic).

CryoGrid.Physics.Soils.SoilParameterization — Type

SoilParameterization

Abstract base type for parameterizations of soil properties.

CryoGrid.Physics.Soils.Westermann — Type

Westermann <: SFCCFunction

Westermann, S., Boike, J., Langer, M., Schuler, T. V., and Etzelmüller, B.: Modeling the impact of wintertime rain events on the thermal regime of permafrost, The Cryosphere, 5, 945–959, https://doi.org/10.5194/tc-5-945-2011, 2011.

CryoGrid.Physics.Soils.mineral — Method

mineral(soil::Soil, ::Heat, state)
mineral(soil::Soil, ::Heat, state, i)

Retrieves the mineral content for the given layer at grid cell i, if provided. Defaults to using the scalar mineral content defined on soil.

CryoGrid.Physics.Soils.organic — Method

organic(soil::Soil, ::Heat, state)
organic(soil::Soil, ::Heat, state, i)

Retrieves the organic content for the given layer at grid cell i, if provided. Defaults to using the scalar organic content defined on soil.

CryoGrid.Physics.Soils.porosity — Method

porosity(soil::Soil, ::Heat, state)
porosity(soil::Soil, ::Heat, state, i)

Retrieves the porosity for the given layer at grid cell i, if provided. Defaults to using the scalar porosity defined on soil.