Soils

CryoGrid.Physics.Soils.CharacteristicFractions — Type

CharacteristicFractions{P1,P2,P3,P4} <: SoilParameterization

Represents uniform composition of a soil volume in terms of fractions: excess ice, natural porosity, saturation, and organic solid fraction.

CryoGrid.Physics.Soils.DallAmico — Type

DallAmico <: SFCCFunction

Dall'Amico M, 2010. Coupled water and heat transfer in permafrost modeling. Ph.D. Thesis, University of Trento, pp. 43.

CryoGrid.Physics.Soils.McKenzie — Type

McKenzie <: SFCCFunction

McKenzie JM, Voss CI, Siegel DI, 2007. Groundwater flow with energy transport and water-ice phase change: numerical simulations, benchmarks, and application to freezing in peat bogs. Advances in Water Resources, 30(4): 966–983. DOI: 10.1016/j.advwatres.2006.08.008.

CryoGrid.Physics.Soils.SFCC — Type

SFCC{F,S} <: FreezeCurve

Generic representation of the soil freeze characteristic curve. The shape and parameters of the curve are determined by the implementation of SFCCFunction f. Also requires an implementation of SFCCSolver which provides the solution to the non-linear mapping H <–> T.

CryoGrid.Physics.Soils.SFCCNewtonSolver — Type

SFCCNewtonSolver <: SFCCSolver

Specialized implementation of Newton's method with backtracking line search for resolving the energy conservation law, H = TC + Lθ. Attempts to find the root of the corresponding temperature residual: ϵ = T - (H - Lθ(T)) / C(θ(T)) and uses backtracking to avoid jumping over the solution. This prevents convergence issues that arise due to discontinuities and strong non-linearity in most common soil freeze curves.

CryoGrid.Physics.Soils.SFCCPreSolver — Type

SFCCPreSolver{TCache} <: SFCCSolver

A fast SFCC "solver" which pre-builds an interpolant for the freeze curve in terms of enthalpy, θ(H). Note that this solver is only valid when all freeze curve parameters are held constant and will produce incorrect results otherwise.

CryoGrid.Physics.Soils.SWRC — Type

SWRC{F}

Soil water retention curve with function type F.

CryoGrid.Physics.Soils.Soil — Type

Basic Soil layer.

CryoGrid.Physics.Soils.SoilParameterization — Type

SoilParameterization

Abstract base type for parameterizations of soil properties.

CryoGrid.Physics.Soils.VanGenuchten — Type

VanGenuchten <: SWRCFunction

van Genuchten MT, 1980. A closed-form equation for predicting the hydraulic conductivity of unsaturated soils. Soil Science Society of America Journal, 44(5): 892–898. DOI: 10.2136/sssaj 1980.03615995004400050002x.

CryoGrid.Physics.Soils.Westermann — Type

Westermann <: SFCCFunction

Westermann, S., Boike, J., Langer, M., Schuler, T. V., and Etzelmüller, B.: Modeling the impact of wintertime rain events on the thermal regime of permafrost, The Cryosphere, 5, 945–959, https://doi.org/10.5194/tc-5-945-2011, 2011.

CryoGrid.Physics.Soils.mineral — Method

mineral(soil::Soil, state, i)

Retrieves the mineral content for the given layer at grid cell i, if provided. Defaults to using the scalar mineral content defined on soil.

CryoGrid.Physics.Soils.organic — Method

organic(soil::Soil, state, i)

Retrieves the organic content for the given layer at grid cell i, if provided. Defaults to using the scalar organic content defined on soil.

CryoGrid.Physics.Soils.porosity — Method

porosity(soil::Soil, state, i)

Retrieves the porosity for the given layer at grid cell i, if provided. Defaults to using the scalar porosity defined on soil.