SciML/DiffEq solvers

CryoGrid.DiffEq.CryoGridEnsembleSetup — Type

CryoGridEnsembleSetup{TTile<:Tile,Tkwargs}

Stores the basic model configuration for a CryoGrid single-tile ensemble.

See also CryoGridEnsembleProblem, fitekp!

CryoGrid.DiffEq.CryoGridEnsembleProblem — Function

CryoGridEnsembleProblem(
    setup::CryoGridEnsembleSetup,
    Θ::AbstractMatrix,
    prob::DiffEqBase.DEProblem=default_initial_problem(setup);
    output_dir=".",
    prob_func=default_ensemble_prob_func(setup, Θ),
    output_func=(sol,i) -> CryoGridOutput(sol),
    reduction=(u,data,i) -> (append!(u,data),false),
    ensprob_kwargs...
)

Constructs an EnsembleProblem from the given ensemble setup for a m x N parameter matrix Θ, where N is the size of the ensemble and m is the dimensionality of the ensmeble state space (e.g. the number of parameters). param_map must be a ParameterMapping with a transform function that accepts an m-dimensional vector and produces a parameter vector (or CryoGridParams instance) which matches the output of CryoGrid.parameters. By default, param_map is the identity function; however, it may be customized to permit the construction of reduced-rank or reparameterized ensembles for which the parameter space differs from the full CryoGrid model parameter space.

Keyword arguments:

output_func: a function (sol,i)::Any which processes the ODESolution for ensemble member i and returns the result. It is recommended to save output to disk for non-trivial time spans to avoid slowdowns from serialization time when running the ensemble using parallel workers.

reduction: a function (u,data,i) which accumulates the result of output_func in u. Defaults to just appending data to u.

output_dir: Only used to specify the output directory for the default implementation of output_func. If a custom output_func is provided, this arugment is ignored.

All additional keyword arguments will be passed to EnsembleProblem.

See also SciMLBase.EnsembleProblem, CryoGridEnsembleSetup, fitekp!