Top level methods and types

CryoGrid.Bottom — Type

Bottom <: Layer

Generic "bottom" layer that marks the lower boundary of the subsurface grid.

CryoGrid.BoundaryProcess — Type

BoundaryProcess{T<:SubSurfaceProcess}

Abstract base type for boundary processes, i.e. processes that operate at the boundaries of the subsurface. A BoundaryProcess represents the boundary conditions of one or more SubSurfaceProcesses but may include its own diagnostic (or even prognostic) variables, if necessary.

CryoGrid.Coupled2 — Type

Coupled2{P1,P2} = CoupledProcesses{Tuple{T1,T2}} where {T1,T2}

Represents an explicitly coupled pair of processes. Alias for CoupledProcesses{Tuple{P1,P2}}. Coupled provides a simple mechanism for defining new behaviors on multi-processes systems.

CryoGrid.CoupledProcesses — Type

CoupledProcesses{TProcs} <: Process

Represents an explicitly or implicitly coupled system of processes. TProcs is always a Tuple of other processes.

CryoGrid.Layer — Type

Abstract base type for all layers.

CryoGrid.Process — Type

Abstract base type for all dynamical processes.

CryoGrid.SubSurface — Type

SubSurface <: Layer

Abstract base type for layers in the stratigraphy, e.g. soil, snow, pond, etc.

CryoGrid.SubSurfaceProcess — Type

SubSurfaceProcess <: Process

Abstract base type for subsurface processes, i.e. processes that operate at or below the surface, such as heat conduction, water infiltration, etc.

CryoGrid.Top — Type

Top <: Layer

Generic "top" layer that marks the upper boundary of the subsurface grid.

CryoGrid.Coupled — Method

Coupled(ps::Process...)

Alias for CoupledProcesses(ps...).

CryoGrid.boundaryflux — Method

boundaryflux(bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, sbc, ssub)
boundaryflux(s::BoundaryStyle, bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, sbc, ssub)

Computes the flux dH/dt at the boundary layer. Calls boundaryflux(BoundaryStyle(B),...) to allow for generic implementations by boundary condition type.

CryoGrid.boundaryvalue — Method

boundaryvalue(bc::BoundaryProcess, lbc::Union{Top,Bottom}, proc::SubSurfaceProcess, lsub::SubSurfaceProcess, sbc, ssub)

Computes the value of the boundary condition specified by bc for the given layer/process combinations.

CryoGrid.criterion — Method

criterion(::Event, ::Layer, ::Process, state)

Event criterion/condition. Should return a Bool for discrete events and a real number for continuous events.

CryoGrid.diagnosticstep! — Method

diagnosticstep!(l::Layer, p::Process, state)
diagnosticstep!(l::Layer, state)

Defines the diagnostic update for a Process on a given Layer.

CryoGrid.diagnosticstep! — Method

diagnosticstep!(l::Layer, ps::CoupledProcesses{P}, state) where {P}

Default implementation of diagnosticstep! for coupled process types. Calls each process in sequence.

CryoGrid.initialcondition! — Method

initialcondition!(::Layer, state)
initialcondition!(::Process, state) = nothing
initialcondition!(::Layer, ::Process, state)

Defines the initial condition for a given Process and/or Layer. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.initialcondition! — Method

initialcondition!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines the initial condition for two processes on adjacent layers. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.initialcondition! — Method

initialcondition!(l1::Layer, ps1::CoupledProcesses{P1}, l2::Layer, ps2::CoupledProcesses{P2}, s1, s2) where {P1,P2}

Default implementation of initialcondition! for coupled process types. Calls each process in sequence.

CryoGrid.initialcondition! — Method

initialcondition!([l::Layer,] ps::CoupledProcesses{P}, state) where {P}

Default implementation of initialcondition! for coupled process types. Calls each process in sequence.

CryoGrid.interact! — Method

interact!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines a boundary interaction between two processes on adjacent layers. For any interaction, the order of the arguments follows decreasing depth, i.e. the first layer/process is always on top of the second layer/process. This ordering matters and separate dispatches must be provided for interactions in reverse order.

CryoGrid.interact! — Method

interact!(l1::Layer, ps1::CoupledProcesses{P1}, l2::Layer, ps2::CoupledProcesses{P2}, s1, s2) where {P1,P2}

Default implementation of interact! for coupled process (CoupledProcesses) types. Generates a specialized implementation that calls interact! on all pairs of processes between the two layers. Since it is a generated function, the process matching occurs at compile-time and the emitted code will simply be a sequence of interact! calls. Pairs of processes which lack a definition of interact! should be automatically omitted by the compiler.

CryoGrid.observe — Method

observe(::Val{name}, l::Layer, ps::CoupledProcesses{P}, state) where {P}

Default implementation of observe for coupled process types. Calls each process in sequence.

CryoGrid.observe — Method

observe(::Val{name}, ::Layer, ::Process, state1)

Called at the end of each step. Can be used by the user to add additional observables via @log without affecting the model implementation. As such, this function should not be used in implementations, but only by users and code which monitors/consumes CryoGrid model outputs.

Example:

observe(::Val{:meanT}, ::SubSurface, ::Heat, state) = @log meanT = mean(state.T)
# build model
...
setup = Tile(stratigraphy, grid, observed=[:meanT])
# solve
...
# retrieve results
@show out.log.meanT # will be a DimArray of meanT at each timestep.

CryoGrid.prognosticstep! — Method

prognosticstep!(l::Layer, p::Process, state)

Defines the prognostic update for a Process on a given layer. Note that an instance of prognosticstep! must be provided for all non-boundary (subsurface) processes/layers.

CryoGrid.prognosticstep! — Method

prognosticstep!(l::Layer, ps::CoupledProcesses{P}, state) where {P}

Default implementation of prognosticstep! for coupled process types. Calls each process in sequence.

CryoGrid.timestep — Method

timestep(::Layer, ::Process, state)

Retrieves the recommended timestep for the given Process defined on the given Layer. The default implementation returns Inf which indicates no timestep restriction. The actual chosen timestep will depend on the integrator being used and other user configuration options.

CryoGrid.timestep — Method

timestep(l::Layer, ps::CoupledProcesses{P}, state) where {P}

Default implementation of timestep for coupled process types. Calls each process in sequence.

CryoGrid.trigger! — Method

trigger!(::Event, ::Layer, ::Process, state)
trigger!(ev::ContinuousEvent, ::ContinuousTrigger, ::Layer, ::Process, state) where {name}

Event action executed when criterion is met.

CryoGrid.variables — Method

variables(::Layer)
variables(::Process)
variables(layer::Layer, process::Process)

Defines variables for a given Process and/or Layer. Implementations should return a Tuple of Vars.