Top level methods and types

CryoGrid.Bottom — Type

Bottom{TProc<:BoundaryProcesses} <: Layer

Generic "bottom" layer that marks the lower boundary of the subsurface grid.

CryoGrid.BoundaryCondition — Type

BoundaryCondition

Trait that specifies the kind of boundary condition. This can be used to write generic implementations of interact! that are (relatively) agnostic to specific implementations of BoundaryProcess. A good example of this can be found in the boundaryflux method interface.

CryoGrid.BoundaryCondition — Method

BoundaryCondition(::Type{T})

Can be overriden by BoundaryProcess types to indicate the type of boundary condition, e.g:

BoundaryCondition(::Type{BP}) = Dirichlet()

where BP is a BoundaryProcess that provides the boundary conditions.

CryoGrid.BoundaryProcess — Type

BoundaryProcess{T<:SubSurfaceProcess}

Abstract base type for boundary processes, i.e. processes that operate at the boundaries of the subsurface. A BoundaryProcess represents the boundary conditions of one or more SubSurfaceProcesses but may include its own diagnostic (or even prognostic) variables, if necessary.

CryoGrid.ConstantBC — Type

ConstantBC{P,S,T} <: BoundaryProcess{P}

Constant boundary condition (of any type/unit) specified by value.

CryoGrid.Coupled2 — Type

Coupled2{P1,P2} = CoupledProcesses{Tuple{T1,T2}} where {T1,T2}

Type alias for coupled processes, i.e. CoupledProcesses{Tuple{P1,P2}}. Coupled provides a simple mechanism for defining new behaviors on multi-processes systems.

CryoGrid.CoupledProcesses — Type

CoupledProcesses{TProcs} <: Process

Represents an explicitly or implicitly coupled system of processes. TProcs is always a Tuple of other processes.

CryoGrid.CryoGridProblem — Type

CryoGridProblem{iip,Tu,Tt,Tp,TT,Tcb,Tdf,Tkw} <: SciMLBase.AbstractODEProblem{Tu,Tt,iip}

Represents a CryoGrid discretized PDE forward model configuration using the SciMLBase/DiffEqBase problem interface.

CryoGrid.CryoGridProblem — Type

CryoGridProblem(
    tile::Tile,
    u0::ComponentVector,
    tspan::NTuple{2,Float64},
    p=nothing;
    saveat=3600.0,
    savevars=(),
    save_everystep=false,
    save_start=true,
    save_end=true,
    step_limiter=CryoGrid.timestep,
    safety_factor=1,
    max_step=true,
    callback=nothing,
    isoutofdomain=Strat.domain(tile),
    specialization=SciMLBase.AutoSpecialize,
    function_kwargs=(),
    prob_kwargs...
)

Constructor for CryoGridProblem that automatically generates necessary callbacks for saving diagnostic state variables

CryoGrid.CryoGridProblem — Method

CryoGridProblem(tile::Tile, u0::ComponentVector, tspan::NTuple{2,DateTime}, args...;kwargs...)

CryoGrid.Layer — Type

Layer

Abstract base type for all layers.

CryoGrid.PeriodicBC — Type

PeriodicBC{P,S,T1,T2,T3,T4} <: BoundaryProcess{P}

Periodic boundary condition (of any type/unit) specified by period, amplitude, and phaseshift.

CryoGrid.Process — Type

Abstract base type for all dynamical processes.

CryoGrid.SubSurface — Type

SubSurface <: Layer

Abstract base type for layers in the stratigraphy, e.g. soil, snow, pond, etc.

CryoGrid.SubSurfaceProcess — Type

SubSurfaceProcess <: Process

Abstract base type for subsurface processes, i.e. processes that operate at or below the surface, such as heat conduction, water infiltration, etc.

CryoGrid.Top — Type

Top{TProc<:BoundaryProcesses} <: Layer

Generic "top" layer that marks the upper boundary of the subsurface grid.

CryoGrid.Coupled — Method

Coupled(ps::Process...)

Constructs a composite/coupled process from one or more processes. Alias for CoupledProcesses(ps...).

CryoGrid.Coupled — Method

Coupled(types::Type{<:Process}...)

Convenince method which constructs a CoupledProcesses type corresponding to each type in types, e.g:

Coupled(SnowMassBalance, HeatBalance) = CoupledProcesses{Tuple{T1,T2}} where {T1<:SnowMassBalance, T2<:HeatBalance}

also equivalent to Coupled2{<:SnowMassBalance,<:HeatBalance}.

CryoGrid.basevariables — Method

basevariables(layer::Layer, process::Process)
basevariables(::Layer)
basevariables(::Any)

Defines "base" or common variables for a given Layer, Process, or arbitrary user-defined type. This should be used to define mandatory or shared variables that should not be overridden by subtypes. As such, basevariables should generally only be defined once per type hierarchy (aside from the default definitions) and only on abstract or union types. Implementations should return a Tuple of Vars.

CryoGrid.boundaryflux — Method

boundaryflux(bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, sbc, ssub)
boundaryflux(s::BoundaryCondition, bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, sbc, ssub)

Computes the flux dH/dt at the boundary layer. Calls boundaryflux(BoundaryCondition(B),...) to allow for generic implementations by boundary condition type. Note that this method uses a different argument order convention than interact!. This is intended to faciliate stratigraphy independent implementations of certain boundary conditions (e.g. a simple Dirichlet boundary could be applied in the same manner to both the upper and lower boundary).

CryoGrid.boundaryvalue — Method

boundaryvalue(bc::BoundaryProcess, lbc::Union{Top,Bottom}, proc::SubSurfaceProcess, lsub::SubSurfaceProcess, sbc, ssub)

Computes the value of the boundary condition specified by bc for the given layer/process combinations. Note that this method uses a different argument order convention than interact!. This is intended to faciliate stratigraphy independent implementations of certain boundary conditions (e.g. a simple Dirichlet boundary could be applied in the same manner to both the upper and lower boundary).

CryoGrid.criterion! — Method

criterion!(out::AbstractArray, ev::GridContinuousEvent, ::Layer, ::Process, state)

Event criterion for on-grid (i.e. multi-valued) continuous events. The condition for each grid cell should be stored in out.

CryoGrid.criterion — Method

criterion(::Event, ::Layer, ::Process, state)

Event criterion/condition. Should return a Bool for discrete events. For continuous events, this should be a real-valued function where the event is fired at the zeros/roots.

CryoGrid.diagnosticstep! — Method

diagnosticstep!(l::Layer, state)
diagnosticstep!(l::Layer, p::Process, state)

Defines the diagnostic update for a Process on a given Layer.

CryoGrid.hasfixedvolume — Method

hasfixedvolume(::Type{<:Layer})

Returns true if the given type has fixed/constant volume or false otherwise.

CryoGrid.initialcondition! — Method

initialcondition!(::Layer, state)
initialcondition!(::Layer, ::Process, state)
initialcondition!(::Layer, ::Process, state, initializer)

Defines the initial condition for a given Process and/or Layer. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.initialcondition! — Method

initialcondition!(layer1::Layer, layer2::Layer, state1, state2)
initialcondition!(::Layer, ::Process, ::Layer, ::Process, state1, state2)
initialcondition!(::Layer, ::Process, ::Layer, ::Process, state1, state2, initializer)

Defines the initial condition for two processes on adjacent layers. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.initialcondition! — Method

initialcondition!(l1::Layer, ps1::CoupledProcesses{P1}, l2::Layer, ps2::CoupledProcesses{P2}, s1, s2) where {P1,P2}

Default implementation of initialcondition! for coupled process types. Calls each process in sequence.

CryoGrid.interact! — Method

interact!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines a boundary interaction between two processes on adjacent layers. For any interaction, the order of the arguments follows decreasing depth, i.e. the first layer/process is always on top of the second layer/process. This ordering matters and separate dispatches must be provided for interactions in reverse order.

CryoGrid.interact! — Method

interact!(l1::Layer, ps1::CoupledProcesses{P1}, l2::Layer, ps2::CoupledProcesses{P2}, s1, s2) where {P1,P2}

Default implementation of interact! for coupled process (CoupledProcesses) types. Generates a specialized implementation that calls interact! on all pairs of processes between the two layers. Since it is a generated function, the process matching occurs at compile-time and the emitted code will simply be a sequence of interact! calls. Pairs of processes which lack a definition of interact! will be automatically omitted by the compiler.

CryoGrid.processes — Method

processes(l::Layer)

Fetches the process(es) attached to this layer, if any. Returned value must be of type Process. If the layer has more than one process, they should be combined together with Coupled(procs...).

CryoGrid.prognosticstep! — Method

prognosticstep!(l::Layer, p::Process, state)

Defines the prognostic update for a Process on a given layer. Note that an instance of prognosticstep! must be provided for all non-boundary (subsurface) processes/layers.

CryoGrid.timestep — Method

timestep(::Layer, ::Process, state)

Retrieves the recommended timestep for the given Process defined on the given Layer. The default implementation returns Inf which indicates no timestep restriction. The actual chosen timestep will depend on the integrator being used and other user configuration options.

CryoGrid.timestep — Method

timestep(l::Layer, ps::CoupledProcesses{P}, state) where {P}

Default implementation of timestep for coupled process types. Calls each process in sequence.

CryoGrid.trigger! — Method

trigger!(::Event, ::Layer, ::Process, state)
trigger!(ev::ContinuousEvent, ::ContinuousTrigger, ::Layer, ::Process, state)
trigger!(ev::GridContinuousEvent, ::ContinuousTrigger, ::Layer, ::Process, state)

Event action executed when criterion is met.

CryoGrid.variables — Method

variables(layer::Layer, process::Process)
variables(::Layer)
variables(::Any)

Defines variables for a given Layer, Process, or arbitrary user-defined type. Implementations should return a Tuple of Vars.

CryoGrid.volumetricfractions — Method

volumetricfractions(::SubSurface, state)
volumetricfractions(::SubSurface, state, i)

Get the volumetric fractions of each constituent in the volume (at grid cell i, if specificed). All implementations of volumetricfractions are expected to obey a semi-consistent order in the returned Tuple of fractions; the first three consituents should always be θw,θi,θa, i.e. water, ice, and air, followed by any number of additional constituents which may be defined by the specific layer. There is no feasible way to verify that client code actually obeys this ordering, so be sure to double check your implementation, otherwise this can cause very subtle bugs!