Top level methods and types

CryoGrid.Bottom — Type

Bottom <: Layer

Generic "bottom" layer that marks the lower boundary of the subsurface grid.

CryoGrid.BoundaryProcess — Type

BoundaryProcess

Abstract base type for boundary processes, i.e. processes that operate at the boundaries of the subsurface. A BoundaryProcess represents the boundary conditions of one or more SubSurfaceProcesses but may include its own diagnostic (or even prognostic) variables, if necessary.

CryoGrid.BoundaryStyle — Type

Trait that specifies the "style" or kind of boundary condition. This can be used to write generic implementations of interact! that are (relatively) agnostic to specific implementations of BoundaryProcess. A good example of this can be found in the boundaryflux method interface.

CryoGrid.BoundaryStyle — Method

BoundaryStyle(::Type{T})

Can be overriden by BoundaryProcess types to indicate the type of boundary condition, e.g:

BoundaryStyle(::Type{BP}) = Dirichlet()

where BP is a BoundaryProcess that provides the boundary conditions.

CryoGrid.Callback — Type

Callback

Base type for callback implementations.

CryoGrid.CallbackStyle — Type

CallbackStyle

Trait for callback types.

CryoGrid.CallbackStyle — Method

CallbackStyle(::C)
CallbackStyle(::Type{<:Callback})

Trait implementation that defines the "style" or type of the given callback as any subtype of CallbackStyle, for example Discrete or Continuous.

CryoGrid.Coupled — Type

Coupled{P1,P2} = CoupledProcesses{Tuple{T1,T2}} where {T1,T2}

Represents an explicitly coupled pair processes. Alias for CoupledProcesses{Tuple{P1,P2}}. Coupled provides a simple mechanism for defining new behaviors on multi-processes systems.

CryoGrid.Coupled — Method

Coupled(p1,p2)

Alias for CoupledProcesses(p1,p2).

CryoGrid.CoupledProcesses — Type

CoupledProcesses{TProcs} <: Process

Represents an explicitly or implicitly coupled system of processes. TProcs is always a Tuple of other processes.

CryoGrid.Dirichlet — Type

BoundaryStyle instance for Dirichlet boundary conditions.

CryoGrid.Layer — Type

Abstract base type for all layers.

CryoGrid.Neumann — Type

BoundaryStyle instance for Neumann boundary conditions.

CryoGrid.Process — Type

Abstract base type for all dynamical processes.

CryoGrid.SubSurface — Type

SubSurface <: Layer

Abstract base type for layers in the stratigraphy, e.g. soil, snow, pond, etc.

CryoGrid.SubSurfaceProcess — Type

SubSurfaceProcess <: Process

Abstract base type for subsurface processes, i.e. processes that operate at or below the surface, such as heat conduction, water infiltration, etc.

CryoGrid.Top — Type

Top <: Layer

Generic "top" layer that marks the upper boundary of the subsurface grid.

CryoGrid.affect! — Method

affect!(c::Callback, ::Layer, ::Process, state)

Callback action executed when criterion is met (boolean condition for discrete callbacks, zero for continuous callbacks).

CryoGrid.boundaryflux — Method

boundaryflux(bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, s1, s2)
boundaryflux(s::BoundaryStyle, bc::BoundaryProcess, b::Union{Top,Bottom}, p::SubSurfaceProcess, sub::SubSurface, s1, s2)

Computes the flux dH/dt at the boundary layer. Calls boundaryflux(BoundaryStyle(B),...) to allow for generic implementations by boundary condition type.

CryoGrid.boundaryvalue — Method

boundaryvalue(bc::BoundaryProcess, b, p, layer, sbc, ssub)

Computes the value of the boundary condition specified by bc for the given layer/process combinations.

CryoGrid.criterion — Method

criterion(c::Callback, ::Layer, ::Process, state)

Callback criterion/condition. Should return a Bool for discrete callbacks and a real number for continuous callbacks.

CryoGrid.diagnosticstep! — Method

diagnosticstep!(l::Layer, p::Process, state)

Defines the diagnostic update for a Process on a given Layer.

CryoGrid.initialcondition! — Method

initialcondition!(::Layer, state)
initialcondition!(::Process, state) = nothing
initialcondition!(::Layer, ::Process, state)

Defines the initial condition for a given Process and/or Layer. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.initialcondition! — Method

initialcondition!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines the initial condition for two processes on adjacent layers. initialcondition! should write initial values into all relevant state variables in state.

CryoGrid.interact! — Method

interact!(::Layer, ::Process, ::Layer, ::Process, state1, state2)

Defines a boundary interaction between two processes on adjacent layers. For any interaction, the order of the arguments follows decreasing depth, i.e. the first layer/process is always on top of the second layer/process. This ordering matters and separate dispatches must be provided for interactions in reverse order.

CryoGrid.observe — Method

observe(::Val{name}, ::Layer, ::Process, state1)

Called at the end of each step. Can be used by the user to add additional observables via @log without affecting the model implementation. As such, this function should not be used in implementations, but only by users and code which monitors/consumes CryoGrid model outputs.

Example:

observe(::Val{:meanT}, ::SubSurface, ::Heat, state) = @log meanT = mean(state.T)
# build model
...
setup = Tile(stratigraphy, grid, observed=[:meanT])
# solve
...
# retrieve results
@show out.log.meanT # will be a DimArray of meanT at each timestep.

CryoGrid.prognosticstep! — Method

prognosticstep!(l::Layer, p::Process, state)

Defines the prognostic update for a Process on a given layer. Note that an instance of prognosticstep! must be provided for all non-boundary (subsurface) processes/layers.

CryoGrid.variables — Method

variables(::Layer)
variables(::Process)
variables(layer::Layer, process::Process)

Defines variables for a given Process and/or Layer. Implementations should return a Tuple of Vars.