Utilities

CryoGrid.Utils.Params — Type

Base type for composite parameter types. Permits iteration of struct fields.

CryoGrid.Utils.convert_tspan — Method

convert_tspan(tspan::Tuple{DateTime,DateTime})
convert_tspan(tspan::Tuple{Float64,Float64})

Convenience method for converting between Dates.DateTime and solver time.

CryoGrid.Utils.dustrip — Method

Debug ustrip. Remove units if and only if debug mode is NOT enabled.

CryoGrid.Utils.structiterate — Method

Provides implementation of Base.iterate for structs.

CryoGrid.Utils.@Float_str — Macro

Similar to @UT_str but produces a Float64 quantity type for the given unit if and only if debug mode is enabled. If debug mode is not enabled, plain Float64 is used instead.

CryoGrid.Utils.@Real_str — Macro

Similar to @UT_str but produces a Real quantity type for the given unit if and only if debug mode is enabled. If debug mode is not enabled, plain Float64 is used instead.

CryoGrid.Utils.@UFloat_str — Macro

Similar to Unitful.@u_str (i.e. u"kg") but produces the type of the quantity rather than the instance. NOT conditional on debug mode.

CryoGrid.Utils.@UT_str — Macro

Similar to Unitful.@u_str (i.e. u"kg") but produces the type of the unit rather than the instance. NOT conditional on debug mode.

CryoGrid.Utils.@setscalar — Macro

Convenience macro for setting scalar (single-element) arrays/vectors. It turns an expression of the form: a.b = ... into a.b[1] = ...

This is primarily intended for code clarity, i.e to clearly discern scalar and non-scalar values.

CryoGrid.Utils.@xu_str — Macro

Similar to Unitful.@u_str (i.e. u"kg") but conditional on debug mode being enabled. Otherwise, no unit is applied. This should be used to apply units (and thus dimensional analysis checks) to physical quantities at test time but not during normal execution to avoid unnecessary overhead.